split TestHelper class into static TestHelper with utility methods and abstract TestBase to be inherited by test classes
