+++ test/Datadog.Trace.ClrProfiler.IntegrationTests/TupleList.cs
@@ -1,0 +1,13 @@
+using System;
+using System.Collections.Generic;
+
+namespace Datadog.Trace.ClrProfiler.IntegrationTests
+{
+    public class TupleList<T1, T2> : List<Tuple<T1, T2>>
+    {
+        public void Add(T1 item, T2 item2)
+        {
+            Add(new Tuple<T1, T2>(item, item2));
+        }
+    }
+}
